+++ src/canvas/canvas.class.ts
@@ -61,7 +61,7 @@ type TDestroyedCanvas = Omit<
  * @extends StaticCanvas
  * @tutorial {@link http://fabricjs.com/fabric-intro-part-1#canvas}
  *
- * @fires object:modified at the end of a transform or any change when statefull is true
+ * @fires object:modified at the end of a transform
  * @fires object:rotating while an object is being rotated from the control
  * @fires object:scaling while an object is being scaled by controls
  * @fires object:moving while an object is being dragged